Added a basic logging module
The log levels are checked in the configuration which, honestly, doesn't feel
like the nicest way of doing things.

I think this could be improved a fair bit by firstly ditching the 'src'
search (as that's going to happen for every single log message) as it's a
run-time solution to a compile-time problem. Using a hash table (which
we can probably compute at compile-time) based on the file/func pairs would
eliminate the searching.

The first rule of optimising is don't, so it's probably worth measuring how
many instructions a normal log message takes (perhaps one with no rule, and one
that is at the default log level).